% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_target}
\alias{adjust_target}
\title{Adjust the target values to account for various factors.}
\usage{
adjust_target(orig, resolution = 0.25, v_max = 69)
}
\arguments{
\item{orig}{vector of original values.}

\item{resolution}{the resolution of the measuring device.}

\item{v_max}{the maximum allowable value of the device.}
}
\value{
the adjusted values.
}
\description{
Attempts to adjust for various limitations with the target data. These
include: (1) guage resolution is likely 0.1mm or less, (2) values over 69 are
extremely unlikely, etc...
}

