% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{createSubmission}
\alias{createSubmission}
\title{Write the submission file.}
\usage{
createSubmission(d, file_name, zip = TRUE)
}
\arguments{
\item{d}{the data to write}

\item{file_name}{the name of the file.}

\item{zip}{boolean to zip the file or not.}
}
\value{
TRUE
}
\description{
Write the submission file.
}

