% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate}
\usage{
interpolate(v, fill = "extend", ...)
}
\arguments{
\item{v}{a vector of data}
}
\description{
A wrapper around zoo::na.fill to interpolate without errors.
}

