% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{mae}
\alias{mae}
\title{Calculate Mean Absolute Error (MAE).}
\usage{
mae(actual, predicted, ...)
}
\arguments{
\item{actual}{vector of test/actual data.}

\item{predicted}{vector of predicted data.}

\item{...}{additional arguments passed to \code{err}}
}
\value{
numerical.
}
\description{
Calculate Mean Absolute Error (MAE).
}

