% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{marshall_palmer}
\alias{marshall_palmer}
\title{Calculate mm/hr from dbz with Marshall Palmer equation.}
\usage{
marshall_palmer(dbz)
}
\arguments{
\item{dbz}{reflectivity as measured by dbz}
}
\value{
estimated rainfall in mm/hr.
}
\description{
Calculate mm/hr from dbz with Marshall Palmer equation.
}

