% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{model_err}
\alias{model_err}
\title{Calculate generic model error.}
\usage{
model_err(actual, predicted, lambda, weights)
}
\arguments{
\item{actual}{vector of test/actual data.}

\item{predicted}{vector of predicted data.}

\item{lambda}{the power factor}

\item{weights}{weights applied to error function.}
}
\value{
a measure of model error.
}
\description{
Calculate generic model error.
}

