% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{model_marshal_palmer}
\alias{model_marshal_palmer}
\title{Estimate Marshal Palmer Rain Expectations}
\usage{
model_marshal_palmer(ref, minutes_past)
}
\arguments{
\item{ref}{radar reflectivity.}

\item{minutes_past}{minutes past the hour.}
}
\value{
vector of expected rain measurements
}
\description{
Estimate Marshal Palmer Rain Expectations
}

