% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_neighbor}
\alias{nearest_neighbor}
\title{Nearest neighbor of a value.}
\usage{
nearest_neighbor(v, n)
}
\arguments{
\item{v}{the value}

\item{n}{the neighbors}
}
\value{
the value of the nearest neighbor.
}
\description{
Nearest neighbor of a value.
}

