% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot histogram and summary values.}
\usage{
plot_histogram(dat, var)
}
\arguments{
\item{dat}{the data frame.}

\item{var}{the variable to plot.}
}
\value{
a plot object
}
\description{
Plot histogram and summary values.
}

