% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features.R
\name{ref_interp}
\alias{ref_interp}
\title{Reflectivity interpolated}
\usage{
ref_interp(r, ...)
}
\arguments{
\item{r}{a vector of reflectivity values}

\item{...}{additional arguments passed to \code{na.fill}}
}
\description{
Reflectivity interpolated
}

