% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval.R
\name{rmse}
\alias{rmse}
\title{Calculate Root Mean Squared Error (RMSE).}
\usage{
rmse(actual, predicted, ...)
}
\arguments{
\item{actual}{vector of test/actual data.}

\item{predicted}{vector of predicted data.}

\item{...}{additional arguments passed to \code{err}}
}
\value{
numerical.
}
\description{
Calculate Root Mean Squared Error (RMSE).
}

