% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_test}
\alias{sample_test}
\title{Sample the data repeatedly and analyze.}
\usage{
sample_test(x, id_var, target_var, FUN, ..., nrounds = 1,
  sample_percent = 0.75)
}
\arguments{
\item{x}{the vector or data frame to sample.}
}
\value{
results of analysis
}
\description{
Sample the data repeatedly and analyze.
}

