% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{select_group}
\alias{select_group}
\title{Select a group of ids from a large data file.}
\usage{
select_group(dat, group_var, s, columns, num_chunks, chunk_size = 10000)
}
\arguments{
\item{dat}{the data.}

\item{group_var}{the group variable.}

\item{s}{vector of group identifiers to be selected.}

\item{columns}{the columns to return, or NULL for all.}

\item{num_chunks}{number of chunks to use.}

\item{chunk_size}{number of rows in each chunk. Overridden by \code{num_chunks}}
}
\value{
selected data.
}
\description{
Select a group of ids from a large data file.
}

