% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features.R
\name{summarize_over_time}
\alias{summarize_over_time}
\title{Calculate summary over time of a variable.}
\usage{
summarize_over_time(times, values)
}
\arguments{
\item{times}{vector of time measurements.}

\item{values}{values to summarize.}

\item{FUN}{function to use to summarize.}
}
\description{
Calculate summary over time of a variable.
}

