% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{time_difference}
\alias{time_difference}
\title{Calculate the time difference.}
\usage{
time_difference(times, num_per_segment = 60)
}
\arguments{
\item{times}{vector of times to be calculated.}

\item{num_per_segment}{number of time steps per segment (e.g. 60 minutes
per hour)}
}
\value{
vector of time differences, weighted to span the entire segment.
}
\description{
Calculate the time difference.
}

