% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{zip2ff}
\alias{zip2ff}
\title{convert zipped csv files to FFDF format.}
\usage{
zip2ff(file, dir_in, dir_out, dir_sep = "/", chunks = 20, headers = TRUE,
  compression = c("gzip"), ...)
}
\arguments{
\item{dir_in}{directory source files are in.}

\item{dir_out}{directory to save rds files in.}

\item{dir_sep}{separator to join file names to directories.}

\item{chunks}{number of chunks to split the text file into while converting.}

\item{headers}{logical indicating csv has headers.}

\item{compression}{type of compression to use or FALSE if none.}

\item{...}{arguments passed to \code{read.csv()}}

\item{files}{list of file names.}
}
\value{
TRUE
}
\description{
convert zipped csv files to FFDF format.
}

